Use a '=' to seperate the kebind and the command for *map command.
This changed makes the commands format more generic for all commands. From the
point of view of the command, there is no difference between the set command
and the various *map commands. Now we use the following format for all
commands:

[count]command[ param1[=value1]].

This will make the commands easier to use and learn and will make our lives
easier in later implementations. For example would be the parsing of
[count] command[ param1=[value1] param2 param3[=value3]] easy.