Removed hint processings out of the javascript.
Cause of the processings that where know in the c-layer and the javascript
layer too, we got really complicated code. Now the javascript-layer did no
know anything of how the hinted sources should be used by the browser. This
only concern that is kept in the javascript are the different types of
elements that are hintable.