Put the default settings in config.h.
Now the default settings are also put in the config.h so that it's easy to
print all default settings and keybindings to stdout. Added new option to dump
all default settings and keybindings to stdout (-D).

Parsing from default config is only a little slower than the previous
approach, but it more equal to the other configuration stuff.