Simplified history saving.
The url history is written to the end of the history file. The history isn't
read into list if not required. Every time we need the history we read it new
from file, so we would always see item added by another browser instance. If
the browser is closed the history files loaded again, duplicates will be
removed and the list shortened to the max history length and written back to
the file.