Fixed wrong mode if no hints where found.
If hinting was started, but no hint was found, the input box was still
activated. Now we set the hint mode immediately in the hints_create function.
In the previous implementation the hinting mode was set in the command
function that didn't know if a hint where found or not and switched to hinting
mode also if no hint mas found.