Fixed no input leafing on some sites.
On some pages like http://dict.leo.org/?search=programming the input mode
could be leafed, but all further typed chars where still written into the
input form field. This makes the browser unusable if such a page was opened.

Now we observe the webview and input box to also catch the types chars away,
if vimb browser is not in insert mode.

The key events are not observed for the whole window to not apply our
keybindings to the webinspector too, that will break the usability of the
webinspector.