Reduced the modes.
According to the vim editor, the number of mode of vimb browser are reduced.
The previous search mode and hinting mode aren't implemented as main modes
with own keybindings, they are now submodes of the normal and command mode
instead. This allow to use the already defined keybindings also for the search
and hinting mode. For example it's possible to navigate in search mode, that
was not possible before.

The searching is now more like in vim, where all the navigation commands work
in search mode too. Also if the searching is aborted by hitting <esc> the
search query keeps memorized and could be continued by command search-{forward,
backward}.