Simplified the shortcut system.
Don't treat the $0 with a special meaning to be filled by the whole query
string. Instead split the query string into as many parts like the highest
available placeholder + 1. That means if only $0 is used in a shortcut this
will also be replaced by the whole query string, but it could also be only the
first parameter in the case when the shortcut contains also other
placeholders.