Fixed garbage file create on :quit (#22).
On 64Bit systems there was created a garbage file with content 'quit', if the
browser was closed by command :quit in inputbox. This happend because the quit
command was run first and freed the command history file, after that the last
command was tried to be saved in history.

This patch moved the save of command in history before the call of the
command, this sounds a little unusual, but I think this is ok for this case.