Added hinting mode to download links (#32).
There is now a new hinting mode and command :hint-save to download the hinted
links into the configured download directory. This hinting is bound to the
keybinding ';s'.

Also the enums out of the command.c file where moved into the command.h file,
because they would be possible needed from other components that call commands
directly and not from their string representation.