Fixed leaving input mode on load commit (#33).
If a page was loaded and text was typed into the inputbox, the input
mode was left on WEBKIT_LOAD_COMMITED. Now we keep in insert mode also if
the page loads to allow to type in a command into inputbox.

Moved modkey unset to keybind.c where the modkeys are set.