Revert "Removed vb_set_mode from the commands."
This reverts commit 97a0fac8fba6fbb958ef3fe20a583f1913b4d1f9.

The setting of the mode, seems to be an essential part of the commands. For
example are some of the commands also used internally and should therefore
also set the right mode after execution. For example the feedback for yanked
uri by hinting where not shown, because after printing the message the hinting
machanism sill observed the inputbox for changed and started hintmode again
which cleared the yank feedback. The same was also seen on ';s' hint-save
command.