Fixed wrong xpath syntax.
Moved the //a[@href] to the front of the xpath expressions in hope this
will speedup the hint processing a little. Also disallowed to hint
disabled or readonly input fields. Fixed wrong replacements of "$C" in xpath
expressions, which only relaces the first occurrence of the needle in the
xpath string.