Revert "Don't bubble up click events for hints (#17)."
This reverts commit 78bfdc4fdee1e84393f268da6ae4d1c24ef291ed.

This patch cause other none usable javascript triggered links. The previous
problem that should be fixed, does also occur if the links are clicked by
mouse so we should try to fix this issue in another place.

It seems that there is a problem if window.open() is triggered in the sites.