Fixed broken completion for long uri (#39).
If the completion reached a uri that where longer than 255 chars the
completion reduced to one possible item. The reason was the vb_echo* functions
that used only 255 chars to put into inputbox. If the uri was longer we
considered the current text and the text of completion item as different and
cleaned the completion and created a new list with the current text of 255
chars which normally matched only once.

The buffer size is now set to 512 for all internal used string buffer.