Little history precessing improvement.
The history files and items list must be made unique so we generated a list
and removed old items that where equal to a new to insert item. This caused
some deletions in the GList that are not necessary. Now the file lines are
processed from the last to the first. That means, every new item (older item
in file) that is already is in the list can be discarded, because the latest
is already in the GList.