Revert "Show if the proxy could not be set."
This reverts commit 87c48feca23acf755558ef8bb766eff3582a40da.

This causes irritating error messages on startup, because we have enabled
proxy by default. So every time a new window is opened we get the error
message that the proxy could not be set, because of a none set environment
variable. If the user adds 'set proxy=off' to his config file would not cause
vimb from throwing this error, because the default config is processed before
the users config and the error has already be thrown.
At the moment I think, this error message isn't required so we can remove it.