Changed hinting visibility check.
Don't even check if the point in the center of the element is the element
itself, because this caused broken links to be not hintable, if the second
link line didn't reach the center of the above line. So now the element is
checked for the upper left corner too.