Removed the git version stuff.
The git version could lead to cases where the version seems to be wrong, in
the case where no tags are pulled from upstream. So I decided to throw this
away. Bay the was the none posix sed command to convert project name to word
starting with upper case char, was replaces by a awk call.