Changed the way hints are processed in js (#53).
To allow to enable the new hint modes g;{MODE} it's better to filter the hint
texts without using xPath. So we filter only those items that are actual valid
and don't need to query the whole document for matching items.
On the other side, do we have more control about how the pattern matching is
done, case sensitive, matching word beginnings or fuzzy search or what ever.