Added validation for prev- nextpattern settings (#54);
By the way the return type of the setting function was changed from boolean to
enum, to allow special error messages shown to the users. In case the boolean
where used, the calling function generated the error message for the user, a
possible more useful error message from the setting function was overruled by
the generic one.