Reset inputbox style if command mode is started or box cleared.
If there was shown a error message in inputbox and the user starts to enter a
new command or presses <Esc>, the inputbox keeps red colored until the timeout
reset the style back to normal.
This patch uses the vb_echo* functions to put the text into the input box,
that will set the style explicit to 'normal'.