Fixed none removed hints on pages with AJAX (#61).
The load of page parts with AJAX triggered the injection of the VbHint object
into the page. This overlayed the current used VbHint object so that the
clear() call (started out of hints.c) was run on the new and empty VbHint
object, that does not know anything of already active hints.

This patch moves the initializing of hints to "window-object-cleared" signal.
This is only emmited, if a new window object is created, so thatwe do not
overlay an already defined object.

According to the surf browser also the user script loading was moved to the
"window-object-clear" signal handler, else this would be called multiple times
on each page load or if only parts of the page are updated via AJAX.