Fixed wrong download path for pages with trailing /.
If a page has a URL with a trailing slash, the suggested file name was an
empty string. But we appended this empty string as file to the download
directory path, so vimb tries to save the download under a directory path an
not a file.
This patch checks the case where the filename is empty so that the default
download file name can be used.