Fixed none working 'F' hintmode for webkit2.
If vimb is built against webkit2 webkit's sandbox does not allow to open a
element via JavaScript into a new window. So, webkit does nothing if the
hinting fires a click event on an element that has a none empty target
attribute set.
This patch enables the setting to allow JavaScript to open windows without
user interaction as long as the hinting is active.