Don't expand download-path if set.
It is unusual to expand the variables when they are set. This will irritate
the user in case ':set download-path=$HOME' is set but ':set download-path?'
spits out not what the user set before.
So vimb follows the example given by vim and save the path like is was given
and expands it it it's used.