Fixed none working open hints in new window (#74).
There is already a hack to allow the hinting JavaScript to open a page also in
a new window without user interaction (hinting in no user interaction in the
manner of webkit). To allow to open hints into new window the webkit setting
'javascript-can-open-windows-automatically' is enabled. After fireing a hint,
this is set back to it's initial value. But this seemed to be to early, what
means we potentially disabled to open page in new window right before the new
window was created.
This patch moved the resetting of the option into the clear function, that
should be late enough.