Show the completion list after the hight was set.
If the completion list is displayed before we set the hight, the user might
see the list for a really short time covering the whole webview window before
it's resized to fit into the lower third of the view. The patch moved the
display of the completion list after the list is calculated and the first
entry is marked as active.