Fixed g_source_remove call with none existing source.
The api was changed and spits out a critical error if a source does no exists.
This was the case when the timeout was remove by glib automatically after the
timeout function returned false.
Now the timeout function does return true to not remove the source
automatically, instead the resource is removed by vimb when the timeout is
handled in the map_handle_keys function.