Fixed wrong url in status bar if hsts changed uri (#79).
This is a workaround for a missbehaviour in webkit. If the URI is changed
within the request-queued scope of the soup session, the changes are not
reflected if the URI is taken from webkit_web_view_get_uri(). Event if the
page is fully loaded webkit assumes to be on the wrong URI, but the
webinspector shows the right one.