Remove unneeded SHELL_CMD.
The g_spawn_command_line_async starts already a shell for us, so there is no
need to start a subshell from there. This fixes also some wired quoting
issues that appeared if there where single quotes in the given command string
to avoid shell expansion. In this case the variables where expanded because of
the quoting issue.