Don't use g_shell_parse_argv for shortcuts (#88).
This function was really handy, but it does not allow to ignore unmatched
quotes which lead to errors that could not be handled very well for the user.

This patch adds a stripped down to the minimum parser for the shortcut
parameters to allow also unmatched single- or double quotes.