Don't switch to normal mode in load committed (#95).
On pages like translate.google.com where JavaScript changes the location hash
of the page, webkit performs nearly a new load of the page. This caused vimb
to switch from input to normal mode like intended if a form is submitted out
of input mode.

If only the location hash of a loaded page is changed, the used load-status
are WEBKIT_LOAD_COMMITTED and WEBKIT_LOAD_FIRST_VISUALLY_NON_EMPTY_LAYOUT. So
we can't switch to normal mode in load committed.