Fixed wrong showcmd in status bar.
If there where mapping that start with keys that needs more keys and we wait
over the timeoutlen without typing further, the already typed keys where added
to the showcmd in status bar.

For example if ':nnoremap 1gt' is configured and '1' is typed, '1' is written
to showcmd. After timeoutlen '11' is written. If '1g' was typed, '1g1g' was
written after timeoutlen.

This patch adds a counter for keys shown in showcmd, so that we do not add
keys that are already present there.