Fix for WYSIWYG editor on github (#114).
For these editors the click does not reach the editable element, so that vimb
does not switch to insert mode. But the textarea is focused when the click
appears, so that we can pick up the click event and check if the focused
element is editable.
I hope this will not break other pages that heavily depend on JavaScript.