Fixed window resize by to long text in inputbox.
The GTK3 text view grows with it's content and forces the whole main window to
increase. To fix this, the text view is put into a scrolled window. This might
be a hack, because we do not use the scrolled window itself, but this restores
the behaviour of the plain text view of GTK2.