Fixed none clickable search suggestion (#134).
Search suggestions on google where not clickable since the patch for the
WYSIWYG editors. This was caused by a wrong return value in the callback for
button-release-event signal.
Renamed variable for the return value, this was wrong, because TRUE means stop
click propagation and not to propagate like the variable name purports.