Fixed none complete keysequence on "<Esc>.
If " is typed in normal mode vimb expects a register char so that the " is
show in showcmd section of the statubar. All following next none register
chars are appended to the key sequence so that the command is not ready to
performed.
This patch switch back to starting parser state if the char after the " is no
valid register.