Clear inputbox after running command successfully (#104).
If a ex command was run successfully the inputbox is cleared if the command
does not omit this. So the user have a fast response if the command failed or
not. This might be useful for :open command where it could take some time
until the user see if the requested page starts to load or not.