Change main HSTS URI in navigation-policy-decision-requested scope (#146).
In case a URL is rewritten to http for a host that is a known hsts host, the
soup session feature rewrites the uri to https like expected. But webkit does
not honor the changed scheme and port and does still return the http uri in
webkit_web_view_get_uri().
To avoid irritations, the URI of the main request is checked in scope of
navigation-policy-decision-requested event. If the uri must be changed, the
changed uri is used for webkit_web_view_load_uri() to start a new request.