Don't give key presses to hint mode if this is off.
If commands where typed into inputbox, there where javaScript errors written
in case a ';' was typed. The decision when the hinting mode is active was
wrong. Hinting is always started explicit so we can check the mode flag to
decide if the key can be handled by the hinting mode.