Don't save mapped commands in history (#130).
If a map causes vimb to run various ex commands or too do a search it's
irritating when these resolved commands are written into history. So a command
is only be written to history, if it's built from typed chars.
At the time the whole history recording is toggled on and off, which is not
what vim does. Maybe it would be better to allow fine gained control which
types of history and also registers are enabled or not.

Now also the command triggered from remote are recorded into history. This
makes sense, because the remote is assumed as normal user input.