Don't block search and url history if not typed (#133).
If a command was not typed, it was not stored in history. But this is
irritating if there is a map that opens a URI. In this case the URI was not
written ti URL history, which feels wrong. So now only the ex command history
recording is blocked if the content was not typed by the user or where given
via vimb fifo.