Changed option --fifo-name into --fifo.
The fifo with given name could not be used for new spawned vimb instances,
because each instance needs it's own fifo. To not make the things to
complicated the fifo is now always named vimb-fifo-{pid}. This allows to use
the fifo also for new spawned instances.
To get the right fifo path on startup the -d or --dump option was added to
print the full fifo path to stdout.