Don't clear focus if vimb is in input mode.
If strict-focus is enabled and the user started typing into form field right
before the page way fully loaded, the focus was cleared, so that vimb switched
back to normal mode and executed the pressed keys as normal mode commands.
No the focus is only cleared if strict-focus is enabled and vimb is not in
input mode.