Remove input timeout if inputbox was changed meanwhile.
If the user started a command that failed with error state, the input timeout
was started. In case this was followed by a good command that prints output to
inputbox, the output was cleared by the previous set timeout function.
To fix this behaviour, the timeout is removed if the timer was started and the
inputbox is changed.