Use hash table for duplicate check of history (#163).
The previous duplication check where don on the generated list of history
items with callback function which is really slow for large history files. Now
the histories url is put into a hash table for a faster duplicate check.
Additional to this some unneeded memory allocation where removed. This makes
the code a little harder to maintain, but hey we don't want to wast time and
memory here.