Quit hinting on mouse click.
In case the user starts hinting and clicked with the mouse on a link, the new
page was opened. But after going back in back/forward history, the old page
was shown with the hints, but the browser was in normal mode so that the hints
where disfunctional and could only be removed by reloading the page.
Now vimb switches back to normal mode in case the user does a mouse click.