Ellipsize left status bar text at the middle.
The left part of the statusbar shows the current url or the url under mouse
pointer or focused hint. The previous ellipsize removed the right part of
these URLs which might be interesting. So this is changed to show the left and
right part of the URL to the to domain par as well as the last path or query
part.