Don't remove error message if command is not known.
If an unknown command was typed into inputbox the inputbox was still color in
the error color, but the content was removed so that the user is not informed
about the reason for the error. Even more critical was this in case the
'input-autohide' is enabled. In this case the clearing of the input box caused
the hiding of this, so that the user will never see that the command failed.