hsts: reload only the main frame
HSTS in done using requeueing. But due to a bug in webkit, a reload of
the frame could be required in order to display the redirected URI (and
not the original one).

The patch force this reload only for the mainframe (where the correct
URI is need), and not for the subframe (where the HSTS stuff is done by
requeueing).

Original idea from @fanglingsu