Add <Space> as special map key name (#176).
Normally the literal space char ' ' can be used for mappings. In case this is
part of the lhs it can be given escaped by '\'. But this does not work on the
rhs, because there is no evaluation of escaping for space like in vim. On the
other hand the rhs starts with the first none whitespace char after the lhs.
So it wasn't possible to apply mapping to sequences that starts with one or
more spaces.
This patch adds the new special key name '<Space>' to be used in this cases.