Fix segfault on config file with invalid commands (#188).
In case the file contained invalid commands, this was shown to the user by
using vb_echo() which writes the error message into the inpubox. But this
triggered the already buffer_changed_cb that delegated the changed buffer
event to the current mode.
Long story short, there was no current mode set before which caused the
segfault.