Allow to escape escape char to in commands (#191).
This patch fixes a misbehaviour in the way the escaping is done for commands
given via inputbox or config file. The previous logic for the lhs of a command
used '\' to escape the ' '. But to get a single '\' it's needed to use '\\'.
But the '\\' was not taken as a single backslash.
Now the backslash escapes ' ' as well as '\'. A backslash followed by any
other char is taken as is.