Check for dom focus/blur event on window object cleared (#201).
This allows us to track also focus changes within frames and iframes also  if
they are loaded dynamically. The previous logic added the event listeners to
the document on WEBKIT_LOAD_FINISHED, but if there where later created iframes
in the dom, these where not observers for focus events.
This is only a first attempt to fix the focus issue and does break the logic
behind `set strict-focus=on`.