Use provisional uri if page could not be loaded (#224).
In case an uri could not be loaded the uri of the webview is still the
previous one. That means the url shown in the url bar is not the user
requested and the 'O' command would also use the wrong url.
To fix this the provisional uri is used for the url bar as well as for the
autocmd call if the load of the uri failed.