Fixed unset of current url on stop loading (#224).
If and url is loaded and the real url is already available in the webview the
provisional data sour seems to be not available anymore. This caused '(null)'
be shown in url bar in case a page loading was aborted by CTRL-C.
To fix this the provisional data source is used to get the current url as long
as it exists, else the webview's url is taken.