More consistent use of title and url (#224).
There where cases where the title and the shown current url did not match the
expectations. For example if a page is opened the right title and url where
shown.
In case a none existing page was opened and another none existing page was
opened the title was set to the failed url. But if the user went back in
history the last failed url was still shown as title.