Allow link activation from search result #131.
When a search is performed and the current highlighted result is part of a
link, a click event is triggered on the link to open it. Currently the click()
is done by JavaScript on the element so that we can't control if the target
open in current window or in a new one.