Allow conditional compilation of xembed
In order to make use of the xembed feature, we require
functionality native to the X11 windowing environment to get a
window's X identifier. Obviously, this requires us to compile
with a Gtk+ library which has support for X, restricting us from
working on Wayland-only platforms.

The issue can be fixed by making the Xembed feature optional. As
it is the only feature requiring platform-specific features, this
neatly allows us to build for a Wayland-only platform when XEMBED
is disabled.