Changed default-zoom behaviour.
In the previous implementation the default-zoom was only a initial full
content zoom applied to the webview. But in case the user changed the
zooming and reset it back by using 'zz' the webkit zoom level was set to
1.0 (100%) and not to the initial zoom level like on startup of vimb.
This behaviour is strange to under stand and does not fit toe the
setting name 'default-zoom'.
To make the default-zoom to a real default, the zoom is also applied in
case the zoom is reseted by the user via 'zz'.