Allow to quit hint mode via <CTRL-C> #349.
There are some pages where the hinting could not be started and the
input box is filled with ';o'. If the user uses <Esc> this key is also
given to the hinting mode which can't handle this so the user has no way
to get back to normal mode.
Moved the <CTRL-C> case to the beginning of the key handler to make sure
the user can abort the current mode by <CTRL-C>.