Inject global hints script during setting.
In case the user disabled user scripts, also the hints script was
removed from the web content manager. So if the page was reloaded the
hinting did not work. To avoid the disabling of vimb internal used
scripts, we set them always during set up of the settings. So they are
readded in case all scripts are removed from the web content manager.