Use direct click method on element #349.
Previously there was some hacky target="_blank" setting unsetting logic
in the hints click method to force to open a hint into new instance or
the same. But this does not work and the hack was previously removed. So
there is no need to do the click event emulation by ourselves, it's
already available as method on the elements itself.