Don't allocate memory for none shown messages.
The logic that decided if a message could be written to input box was in
input_print() which is called with newly allocated and prepared message
string. But in case the widget is focused and we do not intend a forced
update the string preparation was in vain.
So do the check if the input could be written before allocating any
memory.