Fix O after searching caused empty inputbox.
In case a searching is active vimb is still in normal mode. So
'O'-keypress should write ':open URL' to the inputbox and switch vimb
into command mode. But during the switching of the modes, the already
written ':open ..' string was immediately overwritten by empty string.
So the user was left with cursor in empty inputbox.

Now the input is only cleared on stopping searching when a links was
fired during searching which is done in code during
WEBKIT_LOAD_COMMITTED event processing.