Fixed endless loop on hints with same length.
In case hints with same length where generated the loop to get the
number of addressable hints did not end when there was only one single
hintkey defined.
So now do not attempt to run the same length hint label logic in case
there is only one hint key set. In this case normal hint labels are
generated.