Use more async script evaluation #421.
It's better to block UI less by running the hinting stuff. But there are
some points at the time where we can't avoid running the JavaScript
syncronously e.g. to check if the user typed key was handled by the
hinting script or not. Later means that we let the normal mode decide
what to do with this.