Allow extended hinting g-mode also for ;o.
The ';o' hinting was not allowed for extended hinting because it makes no
sense to keep hinting open when opening a link into the same window. But
the ';o' hinting matches also form fields so it is useful to allow it
for extended hinting too. So the user can open the extended hinting by
'g;o' and toggle radio buttons and checkboxes and finally submit the
form.