Switch to input mode on authentication request #444.
Webkit shows an default embedded authentication dialog in case of
HTTP-Authentication. But this dialog is neither a GTK dialog nor part of
the DOM. So we are not informed about the any focus changes and vimb
keeps in normal mode and the user cant use all the chars for the
authentication.
This patch listen to the webview authentication signal to force
switching to input mode. On the other had a new flag is set to avoid
switching vimb back to normal mode because of focus changes in the
underlaying DOM (previous opened page).