history: add history item as soon as the load is committed
When changing websites, a history item for the new website is only being
added as soon as loading it has been finished. This can lead to the
situation where you start on site A, navigate to site B, click on a link
C before the load finished, but land on page A again as soon as you
navigate back.

As the described behavior is really confusing in some cases and is
actively loosing intermediate data, we should add the history item a bit
earlier. To avoid adding calls which immediately redirect to somewhere
else, a good place to add it is the LoadCommitted hook. It fires after
the final redirect as soon as initial data from the site has been
loaded.