Improve scripts change detection in Makefile
When changing a script in `src/scripts`, `make` would need to be invoked
twice for a proper build: once for `scripts.h` and once for the `.o`
files depending on it.

This commit fixes this by integrating `src/scripts/Makefile` into
`src/Makefile` to allow for proper dependency update detection.