Gracefully handle subtleties of util_expand() of `~' and `~user'.
When `~' is expanded to an home directory without any user, g_get_home_dir()
function is used that on Unix systems first try to honor HOME and fall back to
the passwd file if not set.

On the other hand, `~user' is expanded always by checking the passwd
file.  Adjust test_expand_tilde_user() accordingly.